feat: implement changeable tasknames and refactor
